requirements.txt
liac-arff
torchfile
tensorflow>=1.0.0
numpy
scikit-learn